fix delete comment

+++ src/modules/comment/services/CommentServiceImpl.ts
@@ -28,9 +28,10 @@ export class CommentServiceImpl implements ICommentService {
       // kiếm comment
       const comment = await this.commentRepo.findCommentById(data)
       if (!comment) throw new BadRequest('No comment found')
-      // // kiểm tra quyền xóa
-      // const account = await this.accountRepo.findById(comment.account.id)
-      // if (!account) throw new BadRequest('No account found')
+      // kiểm tra quyền xóa
+      const account = await this.accountRepo.findById(comment.account.id)
+      if (!account) throw new BadRequest('No account found')
+      if (account.id !== comment.account.id) throw new BadRequest('You do not have permission to delete this comment')
       // xóa comment
       const response = await this.commentRepo.delete(comment.id)
       if (!response) throw new BadRequest('Delete comment failed')
